Css for blog links
Changed the background from plain black to an image of some of the game characters. I changed the button colours, font and placement to be in the centre rather than at one side.
I created 3 of my own colours in the app.css files for the buttons and the two headers.
I also added my own css for grid displaying the buttons
created own css for the second background image so it didn't move as the page scrolled and it had the correct dimensions